Implemented heat flux between lake and underlying soil.
Previously, the lake bottom was treated as a no flux boundary, so
that no heat traveled between the lake and the underlying soil.  Soil
temperatures in the lake/wetland tile were determined only from the
energy fluxes through the wetland surface.  In other words, the soil
under the lake had the exact same temperature as the soil in the
exposed wetland, as if the lake were not present.

Now, the soil under the lake has its own temperature profile,
independent of that in the exposed wetland.  Heat and any un-absorbed
shortwave radiation in the lowest lake layer now can pass into the top
of the soil column beneath the lake.

There are now two new output variableis that track the soil temperature
profile under the lake:
  OUT_SOIL_TEMP_LAKE:  Average temperature in the hydrologic soil layers beneath the lake
  OUT_SOIL_TNODE_LAKE: Soil thermal node temperatures beneath the lake