Removed global.h and defined global variables in vicNl.c
global.h defined a number of global variables and was only included in
vicNl.c. There is not much of an advantage to this and is likely to lead
to confusion, since all the variables are already declared as externs in
the individual files where they are used or vicNl_def.h.