Merged snow.h with vic_def.h. Removed relic header files.
I combined all the constants into a single vic_def.h file. Mainly to
avoid the problem that the same constants are defined differently in
different parts of the code. As is, there is a risk of "out of sight,
out of mind". Still need to split vic_def.h into a vic_run_def.h and
a vic_driver_def.h