Create vic_run_veg_lib to replace veg_lib
Within vic_run, the global (extern) veg_lib is the one for the
individual grid cell and is therefore not the same as the extern
variable veg_lib used in the driver. For this reason, instances
of veg_lib have been replaced by vic_run_veg_lib in the routines
that are part of vic_run.