Fixed bug in reading state file when skipping cells.
When dist_prec was removed, I neglected to remove a related line in read_initial_model_state.c that assumes (when skipping over inactive cells) that dist_prec parameters occupy 1 line per veg tile in the state file.  This resulted in VIC not being able to read a state file if it had to skip inactive cells.  This has been fixed.