Fix for issues #138 and #145.
In runoff.c, added check to handle case when sum_liq = 0.  Removed print of warning when evap_sum != 0 (this in itself is not an error) and added update of layer[lindex].evap to equal the sum of evap[lindex][frost_area] over all frost_areas after adjustment for avail_liq to ensure consistency.