Fix PET.
This addresses issues #276 and #346.  This replaces 6 output variables OUT_PET_* with a single OUT_PET.  This replaces an internal array of 6 flavors of PET, handled as separate veg tiles, with a single scalar PET per veg tile (and elevation band).  This PET corresponds most closely to the old OUT_PET_NATVEG in that it uses Penman-Monteith with the architectural resistance and LAI of the current vegetation, but it is now an area-weighted sum of potential transpiration from the current vegetation and potential soil evaporation.